apps/api/pyproject.toml
[project]
name = "axelseo-api"
version = "0.1.0"
description = "AxelSEO backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "playwright>=1.49.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "boto3>=1.36.0",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.159.0",
    "structlog>=24.4.0",
    "PyJWT[crypto]>=2.9.0",
    "axelseo-crawler",
    "axelseo-auditor",
    "psycopg2-binary>=2.9.10",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
